perception+gpu: bin columns in float64 too, the other half of frame 7
The float64 elevation fix took frame 7 from 3 differing pixels to 1. The
survivor is the azimuth axis, which the first pass deliberately left alone
because none of the six original points was near a column edge. The one that
remains is:

    pixel (v=39, u=468)   cpu winner 89740   gpu winner 89741
    point 89740 -> column 467 in float64, 468 on the Kaggle host
    azimuth 3.155e-05 of a bin from the edge, about 1.6 float32 ULP

So columns now go through float64 in all three places that compute them:

  - perception/range_image.py: float64 arctan2, and the column arithmetic
    stays float64 rather than being rounded to float32
  - gpu/cuda_kernels.py project_keys: `double az`, and `pi`/`d_theta` arrive
    as float64 instead of float32
  - gpu/visibility.py spherical_project: already float64, unchanged

Rounding the host's azimuth to float32 was tried first and broke
test_columns_match_jp_projection -- visibility's scratch is float64 and it
bins columns in float64, so float64 is the value the two already had to agree
on. Matching that is what makes all three paths consistent rather than two of
three.

775 tests pass, gpu_parity 60/60 on the laptop, and the final map hash is
UNCHANGED at 55c47b58cd3ab52de12991fcbd77df76 -- this does not move any result
on this machine, it only makes the same result portable.

+++ src/gpu/device.py
@@ -221,7 +221,7 @@ class DevicePerception:
         with stage("range_image"):
             self.pts[:4 * n].set(points.reshape(-1))
             self.world[:3 * n].set(np.ascontiguousarray(points_world, np.float64).reshape(-1))
-            K.project_keys()(self.pts, 4, np.float32(np.pi), np.float32(self.d_theta),
+            K.project_keys()(self.pts, 4, np.float64(np.pi), np.float64(self.d_theta),
                              self.phi_max, self.d_phi, self.h, self.w,
                              self.key[:n], self.fov[:n])
             self.key[:n].sort()
